delete old rocketchat file :)
